feat(auth): refactor guards and strategies to common module; update imports across controllers
