+++ be-camp-cuss/src/app.module.ts
@@ -6,11 +6,11 @@ import { ConfigModule } from '@nestjs/config';
 import { StoragesModule } from './storages/storages.module';
 import { OrdersModule } from './orders/orders.module';
 import { OrdersNotificationsModule } from './orders-notifications/orders-notifications.module';
-import { JwtStrategy } from './auth/strategies/jwt.strategy';
+import { JwtStrategy } from './common/strategies/jwt.strategy';
 import { HttpExceptionFilter } from './common/filters/http-exception.filter';
 import { APP_GUARD, Reflector } from '@nestjs/core';
-import { JwtAuthGuard } from './auth/guards/jwt.guard';
-import { RolesGuard } from './auth/guards/roles.guard';
+import { JwtAuthGuard } from './common/guards/jwt.guard';
+import { RolesGuard } from './common/guards/roles.guard';
 import { LoggerModule } from './common/loggers/logger.module';
 import { RequestContextMiddleware } from './common/contexts/request-context.middleware';
 import { AuthService } from './auth/auth.service';
