+++ be-camp-cuss/src/auth/auth.module.ts
@@ -5,7 +5,6 @@ import { PrismaModule } from '../prisma/prisma.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtModule, JwtModuleOptions } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { JwtStrategy } from './strategies/jwt.strategy';
 import { Env } from '../common/constants/env.constant';
 
 @Global()
@@ -29,7 +28,7 @@ import { Env } from '../common/constants/env.constant';
     }),
   ],
   controllers: [AuthController],
-  providers: [AuthService, JwtStrategy],
+  providers: [AuthService],
   exports: [AuthService, JwtModule],
 })
 export class AuthModule {}
