+++ be-camp-cuss/src/orders/orders.controller.ts
@@ -12,8 +12,8 @@ import { OrdersCoreService } from './services/orders-core.service';
 import { OrdersDriverService } from './services/orders-driver.service';
 import { CreateOrderDto } from './dto/create-order.dto';
 import { UpdateOrderDto } from './dto/update-order.dto';
-import { JwtAuthGuard } from '../auth/guards/jwt.guard';
-import { RolesGuard } from '../auth/guards/roles.guard';
+import { JwtAuthGuard } from '../common/guards/jwt.guard';
+import { RolesGuard } from '../common/guards/roles.guard';
 import { Roles } from '../common/decorators/roles.decorator';
 import { Role } from '../common/enums/role.enum';
 import { User } from '../common/decorators/user.decorator';
